refactor(navdrawer): improve samples for navdrawer
