+++ src/app/menus/menus-routes-data.ts
@@ -1,5 +1,4 @@
 export const menusRoutesData = {
-    "navbar": { displayName: "Navbar Overview", parentName: "Navbar" },
     "navbar-sample-1": { displayName: "Navbar Action Button Icon", parentName: "Navbar" },
     "navbar-sample-2": { displayName: "Navbar Custom Action Icon", parentName: "Navbar" },
     "navbar-style": { displayName: "Navbar Style", parentName: "Navbar" }
