+++ src/app/menus/menus-routing.module.ts
@@ -8,13 +8,8 @@ import { NavbarComponent } from "./navbar/navbar.component";
 import { NavDrawerMiniComponent } from "./navdrawer/nav-drawer-mini/nav-drawer-mini.component";
 import { NavDrawerPinComponent } from "./navdrawer/nav-drawer-pin/nav-drawer-pin.component";
 import { NavDrawerSimpleComponent } from "./navdrawer/nav-drawer-simple/nav-drawer-simple.component";
-import { NavdrawerComponent } from "./navdrawer/navdrawer.component";
 
 export const menusRoutes: Routes = [
-    {
-        component: NavdrawerComponent,
-        path: "navigation-drawer"
-    },
     {
         component: NavDrawerSimpleComponent,
         path: "navigation-drawer-simple"
