+++ src/app/menus/navdrawer/nav-drawer-mini/nav-drawer-mini.component.html
@@ -1,15 +1,14 @@
 <div class="content-wrap" igxLayout>
   <igx-nav-drawer id="navigation" [pin]="true" [pinThreshold]="0">
     <ng-template igxDrawer>
-        <span igxDrawerItem [isHeader]="true"> Header </span>
-        <span igxDrawerItem igxRipple>
-            <igx-icon fontSet="material">home</igx-icon>
-            <span>Home</span>
+        <span *ngFor="let item of navItems" igxDrawerItem [active]="item.text === selected" igxRipple (click)="navigate(item)">
+          <igx-icon fontSet="material">{{ item.name }}</igx-icon>
+          <span>{{ item.text }}</span>
         </span>
     </ng-template>
     <ng-template igxDrawerMini>
-        <span igxDrawerItem igxRipple>
-            <igx-icon fontSet="material">home</igx-icon>
+        <span *ngFor="let item of navItems" igxDrawerItem [active]="item.text === selected" igxRipple (click)="navigate(item)">
+            <igx-icon fontSet="material">{{ item.name }}</igx-icon>
         </span>
     </ng-template>
   </igx-nav-drawer>
