+++ src/app/menus/navdrawer/nav-drawer-mini/nav-drawer-mini.component.scss
