+++ src/app/menus/navdrawer/nav-drawer-pin/nav-drawer-pin.component.html
@@ -1,22 +1,23 @@
 <div class="content-wrap" igxLayout>
-  <igx-nav-drawer id="navigation" #drawer [isOpen]="true">
-    <ng-template igxDrawer>
-      <nav>
-        <span igxDrawerItem [isHeader]="true"> Email Account </span>
-        <span igxDrawerItem igxRipple> Inbox </span>
-        <span igxDrawerItem igxRipple [active]="true"> Drafts </span>
-        <span igxDrawerItem igxRipple> Sent </span>
-        <span igxDrawerItem [isHeader]="true"> Folders </span>
-        <span igxDrawerItem igxRipple> Deleted </span>
-        <span igxDrawerItem igxRipple> Archive </span>
-      </nav>
-    </ng-template>
-  </igx-nav-drawer>
+    <igx-nav-drawer id="navigation" #drawer [isOpen]="true" [pin]="true" [pinThreshold]="0">
+      <ng-template igxDrawer>
+        <nav>
+          <span igxDrawerItem [isHeader]="true">Components</span>
 
-  <main igxFlex>
-    <!-- <button (click)="drawer.toggle()"> Menu </button> -->
-    <span igxButton="icon" igxToggleAction="navigation">
-      <igx-icon fontSet="material">menu</igx-icon>
-    </span>
-  </main>
-</div>
+          <span *ngFor="let item of navItems" igxDrawerItem [active]="item.text === selected" igxRipple (click)="navigate(item)">
+            <igx-icon fontSet="material">{{ item.name }}</igx-icon>
+            <span>{{ item.text }}</span>
+          </span>
+        </nav>
+      </ng-template>
+    </igx-nav-drawer>
+
+    <main igxFlex>
+      <!-- <button (click)="drawer.toggle()"> Menu </button> -->
+      <span igxButton="icon" igxToggleAction="navigation">
+        <igx-icon fontSet="material">menu</igx-icon>
+      </span>
+
+      <h5>{{selected}} content.</h5>
+    </main>
+  </div>
