+++ src/app/menus/navdrawer/nav-drawer-pin/nav-drawer-pin.component.scss
@@ -7,7 +7,6 @@
     left: 0;
 }
 
-main {
-    background-color: #f5f5f5;
-    height: 100%;
+main h5 {
+    padding: 10px;
 }
