+++ src/app/menus/navdrawer/nav-drawer-pin/nav-drawer-pin.component.ts
@@ -6,4 +6,15 @@ import { Component } from "@angular/core";
   templateUrl: "./nav-drawer-pin.component.html"
 })
 export class NavDrawerPinComponent  {
+    public navItems = [
+        { name: "account_circle", text: "Avatar" },
+        { name: "error", text: "Badge" },
+        { name: "group_work", text: "Button Group" }
+    ];
+
+    public selected = "Avatar";
+
+    public navigate(item) {
+        this.selected = item.text;
+    }
 }
