+++ src/app/menus/navdrawer/nav-drawer-simple/nav-drawer-simple.component.html
@@ -2,13 +2,12 @@
   <igx-nav-drawer id="navigation" #drawer [isOpen]="true">
     <ng-template igxDrawer>
       <nav>
-        <span igxDrawerItem [isHeader]="true"> Email Account </span>
-        <span igxDrawerItem igxRipple> Inbox </span>
-        <span igxDrawerItem igxRipple [active]="true"> Drafts </span>
-        <span igxDrawerItem igxRipple> Sent </span>
-        <span igxDrawerItem [isHeader]="true"> Folders </span>
-        <span igxDrawerItem igxRipple> Deleted </span>
-        <span igxDrawerItem igxRipple> Archive </span>
+        <span igxDrawerItem [isHeader]="true">Components</span>
+
+        <span *ngFor="let item of navItems" igxDrawerItem [active]="item.text === selected" igxRipple (click)="navigate(item)">
+          <igx-icon fontSet="material">{{ item.name }}</igx-icon>
+          <span>{{ item.text }}</span>
+        </span>
       </nav>
     </ng-template>
   </igx-nav-drawer>
@@ -18,5 +17,7 @@
     <span igxButton="icon" igxToggleAction="navigation">
       <igx-icon fontSet="material">menu</igx-icon>
     </span>
+
+    <h5>{{selected}} content.</h5>
   </main>
 </div>
