+++ src/app/menus/navdrawer/nav-drawer-simple/nav-drawer-simple.component.scss
@@ -3,3 +3,7 @@
     height: 100%;
     display: flex;
 }
+
+main h5 {
+    padding: 10px;
+}
