+++ src/app/menus/navdrawer/nav-drawer-simple/nav-drawer-simple.component.ts
@@ -1,4 +1,5 @@
-import { Component } from "@angular/core";
+import { Component, ViewChild } from "@angular/core";
+import { IgxNavigationDrawerComponent } from "igniteui-angular";
 
 @Component({
   selector: "app-nav-drawer-simple",
@@ -6,4 +7,19 @@ import { Component } from "@angular/core";
   templateUrl: "./nav-drawer-simple.component.html"
 })
 export class NavDrawerSimpleComponent  {
+    public navItems = [
+        { name: "account_circle", text: "Avatar" },
+        { name: "error", text: "Badge" },
+        { name: "group_work", text: "Button Group" }
+    ];
+
+    public selected = "Avatar";
+
+    @ViewChild(IgxNavigationDrawerComponent, { static: true })
+    public drawer: IgxNavigationDrawerComponent;
+
+    public navigate(item) {
+        this.selected = item.text;
+        this.drawer.close();
+    }
 }
