--- src/app/menus/navdrawer/navdrawer.component.scss
@@ -1,38 +1,0 @@
-.main {
-    position: absolute;
-    top: 0;
-    right: 0;
-    bottom: 0;
-    left: 0;
-    overflow: hidden;
-}
-
-main {
-    padding: 1rem 2rem;
-}
-
-main.mini-padding {
-    padding: 1rem calc(1rem + 60px);
-}
-
-/* Sample styling */
-.setting-title {
-    margin-bottom: 0;
-    color: rgba(228, 28, 119, 0.8) /*#e41c77cc*/ ;
-}
-
-.setting-note {
-    font-size: 14px;
-    color: #484848;
-    margin: 4px 0;
-}
-
-.igx-nav-drawer__aside::-webkit-scrollbar {
-    width: 3px;
-    background: #e4e4e4;
-}
-
-.igx-nav-drawer__aside::-webkit-scrollbar-thumb {
-    background: #0375be;
-    border-radius: 0;
-}
